Better...ish??? not really...
Still a failure. It can create new users, access accounts, store information in text files, withdraw, deposit, and check balance. One problem with creating new accounts. The new account has nothing in the file, so an error occurs when the IDE tries to add an integer and the string of "". Tried to fix it using a long if or statement but it still doesn't function properly. )x I feel pretty dumb